Rename Uno to SingleRouter
